requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1705494584803/work
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy @ file:///C:/b/abs_853rfy8zse/croot/contourpy_1700583617587/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fastapi==0.115.5
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.13.1
fonttools @ file:///C:/b/abs_f47gnfqnx0/croot/fonttools_1713551644747/work
fqdn==1.5.1
fsspec==2024.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio @ file:///C:/b/abs_7d9olmtnd2/croot/grpc-split_1716834609196/work
gymnasium==1.0.0a2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.8
imbalanced-learn==0.12.4
importlib_metadata==8.4.0
importlib_resources @ file:///C:/b/abs_e9f8_0_2dq/croot/importlib_resources-suite_1720641112288/work
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_f55mv94vyg/croot/mkl_fft_1725370278455/work
mkl_random @ file:///C:/b/abs_21ydbzdu8d/croot/mkl_random_1725370276095/work
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.3
opencv-python==4.10.0.84
orjson==3.10.12
overrides==7.7.0
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.2
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing @ file:///C:/b/abs_520r19rysg/croot/pyparsing_1725041656021/work
PyQt5==5.15.10
PyQt5-sip @ file:///C:/b/abs_c0pi2mimq3/croot/pyqt-split_1698769125270/work/pyqt_sip
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
ruff==0.8.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
sip @ file:///C:/b/abs_edevan3fce/croot/sip_1698675983372/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.12
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1727435705122/work/tensorboard-2.18.0-py3-none-any.whl#sha256=107ca4821745f73e2aefa02c50ff70a9b694f39f790b11e6f682f7d326745eab
tensorboard_data_server @ file:///D:/bld/tensorboard-data-server_1728639845982/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=9deb955bd534069f774f78b678a52a92acaf7a2c2687e725510bc913f34738e8
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.9
tinycss2==1.3.0
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
tomlkit==0.12.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchcam==0.4.0
torchsummary==1.5.1
torchvision==0.19.1+cu118
tornado @ file:///C:/b/abs_7bua0304mj/croot/tornado_1718740122405/work
tqdm==4.66.5
traitlets==5.14.3
typer==0.13.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
unicodedata2 @ file:///C:/b/abs_b6apldlg7y/croot/unicodedata2_1713212998255/work
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1731097436890/work
widgetsnbextension==4.0.13
zipp @ file:///C:/b/abs_b0beoc27oa/croot/zipp_1704206963359/work
